backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
httpx==0.28.1
numpy==2.4.2
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
scikit-learn==1.8.0
pytz==2024.1
joblib==1.3.2
python-multipart==0.0.6
prisma==0.15.0
APScheduler==3.11.2